+++ CHANGES.md
@@ -2,6 +2,7 @@ Master
 ---------------------------------
 
 - Support deprecation of fields (#53)
+- Support documentation and deprecation of enum values (#54)
 
 0.1.0 2017-05-25
 ---------------------------------